requirements.txt
langchain
langchain_huggingface
langchain_groq
faiss-cpu
dotenv
google-colab
requests
